# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for getblockio Application",
        "title": "Getblockio",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/getblockio/{apikey_id}": {
            "post": {
                "description": "getblockio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "getblockio"
                ],
                "summary": "getblockio",
                "operationId": "getblockio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Apikey Id",
                        "name": "apikey_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResponseCheckBalanceInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "types.ResponseCheckBalanceInfo": {
            "type": "object",
            "properties": {
                "adress": {
                    "type": "string"
                }
            }
        }
    }
}